Intent: Consolidate validator utilities under utils package Operators involved: [Coherence] Affected invariants: [] Key changes: - Move validator helpers into new tnfr.utils.validators module and expose via utils namespace - Update imports across observers, glyph history, dynamics, and tests to target tnfr.utils.validators - Provide tnfr.validators compatibility shim with deprecation warning Expected risks/dissonances: Minor risk of circular imports mitigated with lazy re-export logic Metrics: No change (C(t), Si, νf, phase) Equivalence map: tnfr.validators -> tnfr.utils.validators
